fix: move diagnostics route from /_diag/ingest to /diagnostics/ingest
