+++ app/diagnostics/ingest/IngestDiagnosticsContent.tsx
@@ -4,19 +4,23 @@ import ScrapedSalesSection from './ScrapedSalesSection'
 
 interface IngestRun {
   id: string
-  status: string
+  started_at: string
+  finished_at?: string
+  source: string
+  dry_run: boolean
   fetched_count: number
   new_count: number
   updated_count: number
-  started_at: string
-  finished_at?: string
-  error_message?: string
+  geocode_calls: number
+  cache_hits: number
+  status: string
+  last_error?: string
 }
 
 interface ScrapedSale {
   id: string
   title: string
-  url?: string
+  url: string
   location_text?: string
   posted_at?: string
   first_seen_at?: string
@@ -27,36 +31,24 @@ interface ScrapedSale {
 export default async function IngestDiagnosticsContent() {
   const supabase = createSupabaseServer()
 
-  // Fetch ingestion runs (mock data for now since we don't have a real table)
-  const mockRuns: IngestRun[] = [
-    {
-      id: 'run_1',
-      status: 'completed',
-      fetched_count: 15,
-      new_count: 3,
-      updated_count: 2,
-      started_at: new Date(Date.now() - 3600000).toISOString(),
-      finished_at: new Date(Date.now() - 3500000).toISOString()
-    },
-    {
-      id: 'run_2',
-      status: 'completed',
-      fetched_count: 12,
-      new_count: 1,
-      updated_count: 4,
-      started_at: new Date(Date.now() - 7200000).toISOString(),
-      finished_at: new Date(Date.now() - 7100000).toISOString()
-    },
-    {
-      id: 'run_3',
-      status: 'failed',
-      fetched_count: 0,
-      new_count: 0,
-      updated_count: 0,
-      started_at: new Date(Date.now() - 10800000).toISOString(),
-      error_message: 'Connection timeout'
+  // Fetch ingestion runs from database
+  let runs: IngestRun[] = []
+  try {
+    const { data: runsData, error: runsError } = await supabase
+      .from('ingest_runs')
+      .select('*')
+      .eq('source', 'craigslist')
+      .order('started_at', { ascending: false })
+      .limit(20)
+
+    if (runsError) {
+      console.error('Error fetching ingest runs:', runsError)
+    } else {
+      runs = runsData || []
     }
-  ]
+  } catch (error) {
+    console.error('Error in diagnostics runs fetch:', error)
+  }
 
   // Fetch scraped sales from Craigslist
   let scrapedSales: ScrapedSale[] = []
@@ -72,102 +64,17 @@ export default async function IngestDiagnosticsContent() {
 
     if (error) {
       console.error('Error fetching scraped sales:', error)
-      // If database query fails, provide mock data
-      scrapedSales = [
-        {
-          id: 'mock_1',
-          title: 'Garage Sale - Furniture & Electronics',
-          url: 'https://sfbay.craigslist.org/example1',
-          location_text: 'San Francisco Bay Area',
-          posted_at: new Date(Date.now() - 3600000).toISOString(),
-          first_seen_at: new Date(Date.now() - 3600000).toISOString(),
-          last_seen_at: new Date(Date.now() - 1800000).toISOString(),
-          status: 'published'
-        },
-        {
-          id: 'mock_2',
-          title: 'Moving Sale - Everything Must Go!',
-          url: 'https://sfbay.craigslist.org/example2',
-          location_text: 'Oakland, CA',
-          posted_at: new Date(Date.now() - 7200000).toISOString(),
-          first_seen_at: new Date(Date.now() - 7200000).toISOString(),
-          last_seen_at: new Date(Date.now() - 3600000).toISOString(),
-          status: 'published'
-        },
-        {
-          id: 'mock_3',
-          title: 'Estate Sale - Antiques & Collectibles',
-          url: 'https://sfbay.craigslist.org/example3',
-          location_text: 'Berkeley, CA',
-          posted_at: new Date(Date.now() - 10800000).toISOString(),
-          first_seen_at: new Date(Date.now() - 10800000).toISOString(),
-          last_seen_at: new Date(Date.now() - 5400000).toISOString(),
-          status: 'published'
-        }
-      ]
-      totalCount = 3
     } else {
       scrapedSales = sales || []
       totalCount = count || 0
-      
-      // If no real data, provide mock data
-      if (scrapedSales.length === 0) {
-        scrapedSales = [
-          {
-            id: 'mock_1',
-            title: 'Garage Sale - Furniture & Electronics',
-            url: 'https://sfbay.craigslist.org/example1',
-            location_text: 'San Francisco Bay Area',
-            posted_at: new Date(Date.now() - 3600000).toISOString(),
-            first_seen_at: new Date(Date.now() - 3600000).toISOString(),
-            last_seen_at: new Date(Date.now() - 1800000).toISOString(),
-            status: 'published'
-          },
-          {
-            id: 'mock_2',
-            title: 'Moving Sale - Everything Must Go!',
-            url: 'https://sfbay.craigslist.org/example2',
-            location_text: 'Oakland, CA',
-            posted_at: new Date(Date.now() - 7200000).toISOString(),
-            first_seen_at: new Date(Date.now() - 7200000).toISOString(),
-            last_seen_at: new Date(Date.now() - 3600000).toISOString(),
-            status: 'published'
-          },
-          {
-            id: 'mock_3',
-            title: 'Estate Sale - Antiques & Collectibles',
-            url: 'https://sfbay.craigslist.org/example3',
-            location_text: 'Berkeley, CA',
-            posted_at: new Date(Date.now() - 10800000).toISOString(),
-            first_seen_at: new Date(Date.now() - 10800000).toISOString(),
-            last_seen_at: new Date(Date.now() - 5400000).toISOString(),
-            status: 'published'
-          }
-        ]
-        totalCount = 3
-      }
     }
   } catch (error) {
-    console.error('Error in diagnostics:', error)
-    // Provide mock data on any error
-    scrapedSales = [
-      {
-        id: 'mock_1',
-        title: 'Garage Sale - Furniture & Electronics',
-        url: 'https://sfbay.craigslist.org/example1',
-        location_text: 'San Francisco Bay Area',
-        posted_at: new Date(Date.now() - 3600000).toISOString(),
-        first_seen_at: new Date(Date.now() - 3600000).toISOString(),
-        last_seen_at: new Date(Date.now() - 1800000).toISOString(),
-        status: 'published'
-      }
-    ]
-    totalCount = 1
+    console.error('Error in diagnostics sales fetch:', error)
   }
 
   return (
     <div className="space-y-8">
-      <IngestRunsSection runs={mockRuns} />
+      <IngestRunsSection runs={runs} />
       <ScrapedSalesSection sales={scrapedSales} totalCount={totalCount} />
     </div>
   )
